[python] do not copy image keep the buffer instead
And set image lookup table (LUT) / PySide.QtGui.QImage.setColorTable()
http://qt-project.org/doc/qt-5.1/qtgui/qimage.html#setColorTable

As stated in the doc PySide.QtGui.QImage
does not holds neither copy the buffer.

Constructs an image with the given width , height and format , that uses
an existing memory buffer, data . The width and height must be specified
in pixels, data must be 32-bit aligned, and each scanline of data in the
image must also be 32-bit aligned.

The buffer must remain valid throughout the life of the
PySide.QtGui.QImage . The image does not delete the buffer at
destruction.